Refactors out some common code, internally for now
